refactor: extract shared loadImageBuffer utility for filesystem/cdn fallback
