+++ src/lib/md/compile.ts
@@ -1,4 +1,3 @@
-import fs from "fs"
 import { join } from "path"
 
 import { SerializeOptions } from "next-mdx-remote/dist/types"
@@ -8,7 +7,9 @@ import remarkSlug from "rehype-slug"
 import remarkGfm from "remark-gfm"
 import remarkHeadingId from "remark-heading-id"
 
-import { CONTENT_DIR, CONTENT_PATH, DEPLOY_URL } from "../constants"
+import { loadImageBuffer } from "@/lib/utils/image"
+
+import { CONTENT_DIR, CONTENT_PATH } from "../constants"
 import { Frontmatter, Layout, TocNodeType } from "../types"
 
 import rehypeImg from "@/lib/md/rehypeImg"
@@ -20,28 +21,6 @@ function preprocessMarkdown(content: string) {
   // Replace heading IDs without escaping to escaped version
   // TODO: move to a separate file and test it more
   return content.replace(/^(#{1,6}.*?)\{(#[\w-]+)\}/gm, "$1\\{$2\\}")
-}
-
-/**
- * Load image buffer from filesystem or CDN fallback.
- */
-async function loadHeroImage(imagePath: string): Promise<Buffer | null> {
-  const normalizedPath = imagePath.replace(/^\//, "")
-  const localPath = join("public", normalizedPath)
-
-  // Try local filesystem first
-  if (fs.existsSync(localPath)) {
-    return fs.readFileSync(localPath)
-  }
-
-  // CDN fallback for serverless
-  try {
-    const res = await fetch(`${DEPLOY_URL}/${normalizedPath}`)
-    if (!res.ok) return null
-    return Buffer.from(await res.arrayBuffer())
-  } catch {
-    return null
-  }
 }
 
 export const compile = async ({
@@ -90,7 +69,7 @@ export const compile = async ({
 
   // If the page has a hero image, generate a blurDataURL for it
   if ("image" in frontmatter) {
-    const buffer = await loadHeroImage(frontmatter.image)
+    const buffer = await loadImageBuffer(frontmatter.image)
     if (buffer) {
       const { base64 } = await getPlaiceholder(buffer, { size: 16 })
       frontmatter.blurDataURL = base64
